- Working Reflection, moving towards SOLID for ClassTagHelper and renamed from ClassTagHelperBase.scala
