+++ app/shared/io/helpers/ReflectionHelper.scala
@@ -2,7 +2,7 @@ package shared.io.helpers
 
 import shared.io.helpers.traits._
 
-class ClassTagHelperBaseConcreteImplementation extends ClassTagHelperBase
+class ClassTagHelperConcreteImplementation extends ClassTagHelper
 
 object ReflectionHelper extends ReflectionHelperBase {
   implicit def toNiceObject[T <: AnyRef](x : T) : NiceObject[T] = new NiceObject(x)
