+++ app/shared/io/helpers/traits/ClassTagHelperMembers.scala
@@ -1,0 +1,11 @@
+package shared.io.helpers.traits
+
+import shared.io.helpers.ClassTagHelperConcreteImplementation
+
+trait ClassTagHelperMembers
+  extends ClassTagHelperMemberInfoGetter
+    with ClassTagHelperEventualMembers
+    with ClassTagHelperEventualMembersMap {
+  this : ClassTagHelperConcreteImplementation =>
+
+}
