+++ app/shared/io/helpers/traits/ClassTagHelperMethods.scala
@@ -1,0 +1,20 @@
+package shared.io.helpers.traits
+
+import java.lang.reflect.Method
+
+import shared.com.ortb.enumeration.ReflectionModifier
+import shared.io.helpers.ClassTagHelperConcreteImplementation
+
+import scala.reflect.ClassTag
+
+trait ClassTagHelperMethods
+  extends ClassTagHelperGetMethodsAsMap
+    with ClassTagHelperGetMethodWrappersAsMap {
+  this : ClassTagHelperConcreteImplementation =>
+
+  def getMethodsWithModifier[T](reflectionModifier : ReflectionModifier)(implicit ct : ClassTag[T]) : Array[Method] =
+    getMethods[T].filter(method => reflectionModifier.value == method.getModifiers)
+
+  def getMethods[T](implicit ct : ClassTag[T]) : Array[Method] =
+    getClass[T].getMethods
+}
