+++ app/shared/io/helpers/traits/ClassTagHelperProperties.scala
@@ -1,0 +1,14 @@
+package shared.io.helpers.traits
+
+import shared.com.ortb.manager.traits.DefaultExecutionContextManagerConcreteImplementation
+import shared.io.helpers.ClassTagHelperConcreteImplementation
+
+import scala.concurrent.ExecutionContext
+
+trait ClassTagHelperProperties {
+  this : ClassTagHelperConcreteImplementation =>
+  lazy val executionContextManager = new DefaultExecutionContextManagerConcreteImplementation
+  lazy val defaultMemberPossibility : Int = 30
+
+  implicit override def createDefaultContext() : ExecutionContext = executionContextManager.defaultExecutionContext
+}
